+++ src/types/Filter.ts
@@ -3,4 +3,5 @@ export type Filter = {
   to: string
 	clientId: string|null
 	driverId: string|null
+	origin: string|null
 }